refactor: tests fixtures folder structure
- group by adapters instead of spotify content types
