+++ tests/fixtures/tidal/tidalAuthResponseMock.json
