+++ tests/integration/page.test.ts
@@ -21,7 +21,7 @@ import { app } from '~/index';
 const INDEX_ENDPOINT = 'http://localhost';
 
 const spotifySongHeadResponseMock = await Bun.file(
-  'tests/fixtures/song/spotifyHeadResponseMock.html'
+  'tests/fixtures/spotify/spotifySongHeadResponseMock.html'
 ).text();
 
 describe('Page router', () => {
